changed type for timetracking from int to double

+++ uebung2/simulationmodel/Arrival.java
@@ -1,21 +1,21 @@
 package simulationmodel;
 public class Arrival extends Event {
-	private int execTime;
+	private double execTime;
 
-	public Arrival(int execTime){
+	public Arrival(double execTime){
 		this.execTime=execTime;
 	}
 
 	public void eventExec(simulatorcore.BasicSimulator simulator){
 		System.out.println("Arrival");
 		simulator.getSimulationEntity().setState(simulator.getSimulationEntity().getState()+1);
-		int arrivalTime = simulator.getCurrentSimTime() + (int)(Math.random()*10) + 1;
+		double arrivalTime = simulator.getCurrentSimTime() + (Math.random()*10) + 1;
 		simulator.getEventList().putAway(new Arrival(arrivalTime));
 		if (simulator.getSimulationEntity().getState() == 1) {  //should only occur when jobs was empty
-			simulator.getEventList().putAway(new Departure(arrivalTime + (int)(Math.random()*6 + 1)));
+			simulator.getEventList().putAway(new Departure(arrivalTime + (Math.random()*6 + 1)));
 		}
 	}
-	public int getExecTime(){
+	public double getExecTime(){
 		return execTime;
 	}
