+++ uebung2/simulationmodel/IEvent.java
@@ -1,5 +1,5 @@
 package simulationmodel;
 public interface IEvent {
 	void eventExec(simulatorcore.BasicSimulator simulator);
-	int getExecTime();
+	double getExecTime();
 }
